Add explicit SystemUIDialog dependencies as parameters
This CL adds a new constructor to SystemUIDialog so that its deps can be
passed explicitly, instead of implicitly like it is currently done using
Dependency.get(). This is the first step to removing those calls to
get() and be able to create a SystemUIDialog with mocked dependencies in
(screenshot) tests.

Bug: 219008720
Test: Manual
Change-Id: Ia34072613c82e47b8fc59150703711f6a3c8c8a0